% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_utils.r
\name{condense_spatial_files}
\alias{condense_spatial_files}
\title{Condense Spatial Files}
\usage{
condense_spatial_files(files)
}
\arguments{
\item{files}{Character vector of file names or paths}
}
\value{
Character vector of file names or paths suitable for
  \code{sf::st_read}
}
\description{
Removes shapefile helper files from a list of file names or paths so they can
  be passed to functions like \code{sf::st_read} that expect a single file
  path for shapefiles
}
\examples{

f <- c("data/locs.shp", "data/locs.shx", "data/locs.dbf", "data/locs.prj")

# sf::st_read(f) results in error (the condition has length > 1)

condense_spatial_files(f)

# sf::st_read(condense_spatial_file(f)) will succeed with only one shapefile

# add another shapefile to the file list

f <- c(
  f,
  "data/tracks.shp", "data/tracks.shx", "data/tracks.dbf", "data/tracks.prj"
)
condense_spatial_files(f)

# helper files have been removed, but st_read can't read multiple datasets
# purrr::map_dfr(condense_spatial_files(f), sf::st_read) will do it

# only shapefiles are collapsed, so other files are unchanged

f <- c(
  f,
  "data/herd_unit.geojson", "data/hunt_areas.kml"
)
condense_spatial_files(f)

# so purrr::map_dfr(condense_spatial_files(f), sf::st_read) will work even
  # with a mix of shapefles and other file types

}
