% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_utils.r
\name{count_spatial_files}
\alias{count_spatial_files}
\title{Count Spatial Files}
\usage{
count_spatial_files(files)
}
\arguments{
\item{files}{Character vector of file names or paths}
}
\value{
int The number of unique datasets represented in \code{files}
}
\description{
Counts the number of files in a vector of file names or paths, treating
  each shapefile (made up of several component files) as a single file
}
\examples{

# a single shapefile is stored in at least 4 files on disk with the same
  # basename

f <- c("data/locs.shp", "data/locs.shx", "data/locs.dbf", "data/locs.prj")
count_spatial_files(f)

# [1] 1

# add another shapefile to the file list

f <- c(
  f,
  "data/tracks.shp", "data/tracks.shx", "data/tracks.dbf", "data/tracks.prj"
)
count_spatial_files(f)

# [1] 2

# only shapefiles are collapsed, so other files each count as 1

f <- c(
  f,
  "data/gps.csv", "data/herd_unit.geojson", "data/hunt_areas.kml"
)
count_spatial_files(f)

# [1] 5

}
