% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.r
\name{create_transects}
\alias{create_transects}
\title{create_transects}
\usage{
create_transects(opt_spacing)
}
\arguments{
\item{opt_spacing}{A list produced by \code{optimal_transect_spacing}
containing the spacing in meters and the arguments used to call the
function}
}
\value{
Returns an \code{sf} object with features of type \code{LINESTRING}
}
\description{
Generate transect lines within the specified polygon at a
  certain spacing for final output
}
\keyword{internal}
