% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.r
\name{demo_gen_transects}
\alias{demo_gen_transects}
\title{Demo Transect Generation}
\usage{
demo_gen_transects(
  sampling_frame,
  total_length_km,
  frame_buffer_km = 0.5,
  angle_deg = 0,
  min_length_km = 5,
  max_length_km = 1000,
  break_buffer_km = 0.5,
  min_space_km = 1
)
}
\arguments{
\item{sampling_frame}{\code{sf} object  with the polygon(s) to be surveyed.
\code{sampling_frame} should use a projected CRS with meters as the unit.}

\item{total_length_km}{Total length of transects to generate in kilometers}

\item{frame_buffer_km}{Distance from polygon edges in kilometers at which
transects start or stop. The actual area allowed for placing transects is
\code{sf::st_buffer(sampling_frame, frame_buffer_km * -1000)}.}

\item{angle_deg}{Orientation of the output transects in degrees.}

\item{min_length_km}{Minimum length in kilometers for any individual
transect. Transects shorter than \code{min_length_km} are discarded from
the survey.}

\item{max_length_km}{Maximum length in kilometers for any individual
transect. Transects longer than \code{max_length_km} are split into
multiple transects with space in between.}

\item{break_buffer_km}{Determines the amount of space separating the new
transects when a transect longer than \code{max_length_km} is split. This
argument determines the distance from the original transect centroid to the
endpoints of the new transects, so the resulting transect endpoints will be
\code{break_buffer_km * 2} kilometers apart.}

\item{min_space_km}{Minimal interval in kilometers to consider for transect
spacing. Should be larger than the width of a single transect at a minimum.}
}
\value{
Returns an \code{sf} object identical to that produced by
  \code{generate_transects}
}
\description{
A much more verbose version of \code{generate_transects} meant to demonstrate
  the steps in that function with plots and console messages.
}
\seealso{
generate_transects
}
