% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_kml.r
\name{labeled_kml}
\alias{labeled_kml}
\title{Labeled KML Polygons}
\usage{
labeled_kml(sf, labels, folder, out_file)
}
\arguments{
\item{sf}{\code{sf} object containing polygons}

\item{labels}{column to use for labels (unquoted a la \code{dplyr})}

\item{folder}{name of folder in kml file}

\item{out_file}{file name or path to be written}
}
\value{
character vector of the lines written to the file
}
\description{
Creates a kml file with labels visible at centroids
}
\examples{

library(dplyr)
library(tibble)
library(sf)

tf <- tempfile(fileext = ".kml")

tibble(
  Name = c("59: 1", "59: 2", "59: 3"),
  wkt = c(
    "MULTIPOLYGON (((
      -105.0916 41.96413, -105.0916 41.94263, -105.0627 41.94263,
      -105.0627 41.96413, -105.0916 41.96413
    )))",
    "MULTIPOLYGON (((
      -105.0627 41.96413, -105.0627 41.94263, -105.0338 41.94263,
      -105.0338 41.96413, -105.0627 41.96413
    )))",
    "MULTIPOLYGON (((
      -105.0338 41.96413, -105.0338 41.94263, -105.0048 41.94263,
      -105.0048 41.96413, -105.0338 41.96413
    )))"
  )
) |>
  st_as_sf(wkt = "wkt", crs = 4326) |>
  mutate(label = paste("Unit", Name)) |>
  labeled_kml(label, "Demo", tf)

}
