% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.r
\name{optimal_transect_spacing}
\alias{optimal_transect_spacing}
\title{optimal_transect_spacing}
\usage{
optimal_transect_spacing(
  poly,
  min_length,
  max_length,
  buffer,
  min_space,
  target_length
)
}
\arguments{
\item{poly}{\code{sf} object with the polygon(s) to be surveyed after any
modifications (region exclusion, rotation)}

\item{min_length}{Minimum length in meters for resulting transects}

\item{max_length}{Maximum length in meters for resulting transects -
any lines longer than \code{max_length} are split at their centroids via
\code{break_lines} until none remain}

\item{buffer}{Buffer in meters added to either side of centroid when
separating line segments}

\item{min_space}{Minimum spacing in meters to consider between transects}

\item{target_length}{Target total length in meters of resulting transects}
}
\value{
Returns a list with the optimal spacing, the absolute value of the
  difference between the length requested and the length resulting from the
  optimal spacing, and the arguments passed into the function for reuse in
  \code{create_transects}
}
\description{
Run three optimization stages to determine the ideal spacing
  for transects within a polygon given a target length
}
\keyword{internal}
