% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_geometry.r
\name{rotate_geometry}
\alias{rotate_geometry}
\title{Rotate SF Geometry}
\usage{
rotate_geometry(sf, deg, byrow = FALSE)
}
\arguments{
\item{sf}{\code{sf} object to rotate}

\item{deg}{Numeric - number of degrees to rotate \code{sf} object by}

\item{byrow}{Logical - if \code{byrow} = \code{TRUE} then geometries are each
rotated independently around their own centroid; if \code{byrow} = FALSE
geometries are rotated together around the common centroid}
}
\value{
Returns an \code{sf} object with rotated geometries
}
\description{
Affine transformation modified from sf vignette
  (https://r-spatial.github.io/sf/articles/sf3.html)
}
\examples{

\dontrun{

ha <- wgfd_agol_data("DeerHuntAreas") |>
  sf::st_transform("ESRI:102212")

r1 <- ha |>
  rotate_geometry(45)

plot(sf::st_geometry(r1))

r2 <- ha |>
  rotate_geometry(45, byrow = TRUE)

plot(sf::st_geometry(r2))

}

}
