% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_geometry.r
\name{rotation_matrix}
\alias{rotation_matrix}
\title{rotation_matrix}
\usage{
rotation_matrix(deg)
}
\arguments{
\item{deg}{Numeric - number of degrees to rotate \code{sf} object by}
}
\value{
Returns a matrix that can be multiplied by the geometry column of an
  \code{sf} object to rotate the geometries
}
\description{
Affine transformation modified from sf vignette
  (https://r-spatial.github.io/sf/articles/sf3.html)
}
\keyword{internal}
