% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.r
\name{test_spacing}
\alias{test_spacing}
\title{test_spacing}
\usage{
test_spacing(spacing, poly, min_length, max_length, buffer, target_length)
}
\arguments{
\item{spacing}{Spacing between transects in meters}

\item{poly}{\code{sf} object with the polygon to be surveyed after any
modifications (region exclusion, rotation)}

\item{min_length}{Minimum length in meters for resulting transects}

\item{max_length}{Maximum length in meters for resulting transects - any
lines longer than \code{max_length} are split at their centroids via
\code{break_lines} until none remain}

\item{buffer}{Buffer added to either side of centroid when separating line
segments}

\item{target_length}{Target total length of resulting transects}
}
\value{
Returns the absolute value of the difference in meters between the
  target length and the total length of the generated transects
}
\description{
Generate transects within the specified polygon(s) at a certain
  spacing for optimization
}
\keyword{internal}
