% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgfd_agol.r
\name{wgfd_agol_data}
\alias{wgfd_agol_data}
\title{WGFD AGOL Data}
\usage{
wgfd_agol_data(layer_name, crs = 4326)
}
\arguments{
\item{layer_name}{Name of the layer to download}

\item{crs}{The coordinate reference system for the output \code{sf} object,
see \code{\link[sf]{st_crs}}}
}
\value{
\code{sf} object
}
\description{
Download WGFD data from ArcGIS Online
}
\examples{

sr <- wgfd_agol_data("Antelope_Seasonal_Range")

plot(sr["RANGE"])

}
